+++ packages/core/button/src/styled/ButtonContent.js
@@ -21,6 +21,7 @@ const ButtonContent = (props: Props) => {
   const style = {
     alignItems: getAlignment(props),
     alignSelf: getAlignment(props),
+    display: 'inline-flex',
     flex: '1 1 auto',
     margin: getMargin(props),
     maxWidth: '100%',