+++ packages/core/field-text-area/src/FieldTextArea.js
@@ -54,7 +54,6 @@ type State = {|
 |};
 
 export default class FieldTextArea extends Component<Props, State> {
-  props: Props; // eslint-disable-line react/sort-comp
   input: any; // eslint-disable-line react/sort-comp
 
   static defaultProps = {