+++ packages/elements/mention/src/__tests__/unit/api/PresenceResourceSpec.ts
@@ -1,3 +1,5 @@
+import 'es6-promise/auto'; // 'whatwg-fetch' needs a Promise polyfill
+import 'whatwg-fetch';
 import * as fetchMock from 'fetch-mock/src/client';
 
 import PresenceResource, {