+++ packages/media/media-card/src/index.ts
@@ -21,15 +21,12 @@ export {
 } from './root/cardView';
 
 export * from './root/domain';
-export * from './list';
+export { CardListProps, CardListState } from './list';
+export { default as CardList } from './list/cardListLoader';
+
 export * from './actions';
 export { isUrlPreviewIdentifier } from './utils/identifier';
 // TODO: don't expose this directly https://jira.atlassian.com/browse/FIL-4396
-export {
-  AppCardView,
-  AppCardModel,
-  convertAppCardToSmartCard,
-} from './app_2/AppCardViewV2';
 
 export type CardStatus =
   | 'uploading'