+++ packages/media/media-card/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/media-image
 
+## 40.0.1
+- [patch] Code split media list [d101ce1](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/d101ce1)
+
+## 40.0.0
+- [major] Remove support for ApplicationCard [6e510d8](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/6e510d8)
+
 ## 39.0.1
 - [patch] Fix bug with download binary [71ebe0b](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/71ebe0b)